back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-spectacular==0.27.2
executing==2.0.1
fastjsonschema==2.20.0
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oauthlib==3.2.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.7.0
yarg==0.1.9
